requirements.txt
streamlit
streamlit-authenticator
playwright
langchain
langchain-openai
langchain-google-genai
langchain-community
chromadb
openai
beautifulsoup4
pypdf
python-docx
youtube-transcript-api
python-dotenv
psycopg2-binary
sqlalchemy
watchdog
pandas
openpyxl
